Bug Fixes 20-2-2023
