+++ includes/classes/class-wcj-tcpdf.php
@@ -67,25 +67,26 @@ if ( ! class_exists( 'WCJ_TCPDF' ) ) :
 		 * @todo    (maybe) e.g. "Set font" - `$this->SetFont( 'helvetica', 'I', 8 );`
 		 */
 		public function Header() {
-		 	$invoice_type          = $this->invoice_type;
+			$invoice_type     = $this->invoice_type;
 			$background_image = get_option( 'wcj_invoicing_' . $invoice_type . '_background_image', '' );
 			if ( '' !== $background_image ) {
 				$parse_bkg_image  = wcj_get_option( 'wcj_invoicing_' . $invoice_type . '_background_image_parse', 'yes' );
 				$document_root    = isset( $_SERVER['DOCUMENT_ROOT'] ) ? sanitize_text_field( wp_unslash( $_SERVER['DOCUMENT_ROOT'] ) ) : '';
 				$background_image = 'yes' === ( $parse_bkg_image ) ? $document_root . wp_parse_url( $background_image, PHP_URL_PATH ) : $background_image;
 			}
-	        // get the current page break margin
-	        $bMargin = $this->getBreakMargin();
-	        // get current auto-page-break mode
-	        $auto_page_break = $this->AutoPageBreak;
-	        // disable auto-page-break
-	        $this->SetAutoPageBreak(false, 0);
-	        $this->Image($background_image, 0, 0, 210, 297, '', '', '', false, 300, '', false, false, 0);
-	        // restore auto-page-break status
-	        $this->SetAutoPageBreak($auto_page_break, $bMargin);
-	        // set the starting point for the page content
-	        $this->setPageMark();
-	    }
+			// get the current page break margin.
+			$b_margin = $this->getBreakMargin();
+			// get current auto-page-break mode.
+			$auto_page_break = 'AutoPageBreak';
+			$auto_page_break = $this->$auto_page_break;
+			// disable auto-page-break.
+			$this->SetAutoPageBreak( false, 0 );
+			$this->Image( $background_image, 0, 0, 210, 297, '', '', '', false, 300, '', false, false, 0 );
+			// restore auto-page-break status.
+			$this->SetAutoPageBreak( $auto_page_break, $b_margin );
+			// set the starting point for the page content.
+			$this->setPageMark();
+		}
 	}
 
 endif;
